Rudimentary test of json schema validation
