+++ esim/es2/esim/src/test/resources/hello-world-schema.json
@@ -1,0 +1,7 @@
+{
+  "type": "object",
+  "properties": {
+    "hello": { "type": "string" }
+  },
+  "required": ["hello"]
+}
